+++ chrome/content/button.js
@@ -15,14 +15,14 @@
  */
 /* exported inspect */
 function inspect(object, own = true) {
-  alert('Found: ' + Object.keys(object).length + ' Type: ' + typeof object);
+  alert(`Found: ${Object.keys(object).length} Type: ${typeof object}`);
   for (let key in object) {
     if (own) {
       if (object.hasOwnProperty(key)) {
-        alert(key + ' = ' + object[key]);
+        alert(`${key} = ${object[key]}`);
       }
     } else {
-      alert(key + ' = ' + object[key]);
+      alert(`${key} = ${object[key]}`);
     }
   }
 }
@@ -57,5 +57,35 @@ function deleteReadEmails() {
       treeView.selection.rangedSelect(i, i, true);
     }
   }
-  // goDoCommand('cmd_delete');
+  goDoCommand('cmd_delete');
 }
+
+/**
+ * Defines an event listener that monitors a change in the current folder's message
+ * count and displays the number of read messages in the add-on's toolbar button
+ *
+ */
+let FolderListener = {
+  /**
+   * https://dxr.mozilla.org/comm-central/source/mailnews/base/public/nsIFolderListener.idl#33
+   * void OnItemIntPropertyChanged(in nsIMsgFolder aItem,
+   *                               in nsIAtom aProperty,
+   *                               in long long aOldValue,
+   *                               in long long aNewValue);
+   *
+   */
+  OnItemIntPropertyChanged: function() {
+    /*
+     * Calculate the number of read emails everytime a message is marked as read
+     */
+    let readEmails = countReadEmails();
+    let button = document.getElementById('btn-delete-read-emails');
+    button.label = `Delete Read Emails (${readEmails})`;
+  }
+};
+
+let mailSession = Cc['@mozilla.org/messenger/services/session;1']
+  .getService(Ci.nsIMsgMailSession);
+
+// Register the listener
+mailSession.AddFolderListener(FolderListener, -1);